+++ utils/core/src/serde/byte_writer.rs
@@ -62,6 +62,24 @@ pub trait ByteWriter: Sized {
         self.write_bytes(&value.to_le_bytes());
     }
 
+    /// Writes a usize value in [vint64](https://docs.rs/vint64/latest/vint64/) format into `self`.
+    ///
+    /// # Panics
+    /// Panics if the value could not be written into `self`.
+    fn write_usize(&mut self, value: usize) {
+        let length = encoded_len(value);
+
+        // 9-byte special case
+        if length == 9 {
+            // length byte is zero in this case
+            self.write_u8(0);
+            self.write(value.to_le_bytes());
+        } else {
+            let encoded_bytes = ((value << 1 | 1) << (length - 1)).to_le_bytes();
+            self.write_bytes(&encoded_bytes[..length]);
+        }
+    }
+
     /// Writes a serializable value into `self`.
     ///
     /// # Panics
@@ -82,4 +100,14 @@ impl ByteWriter for Vec<u8> {
     fn write_bytes(&mut self, values: &[u8]) {
         self.extend_from_slice(values);
     }
+}
+
+// HELPER FUNCTIONS
+// ================================================================================================
+
+/// Returns the length of the value in vint64 enсoding.
+pub fn encoded_len(value: usize) -> usize {
+    let zeros = value.leading_zeros() as usize;
+    let len = zeros.saturating_sub(1) / 7;
+    9 - core::cmp::min(len, 8)
 }